influencer_ads/requirements.txt
streamlit>=1.28.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.116.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0

